# services/faux-schoolday/docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Faux Schoolday API
  description: |
    High-fidelity mock implementation of Schoolday platform APIs for teacher evaluation system integration.

    This mock service provides comprehensive simulation of:
    - OAuth 2.0 authentication flows
    - OneRoster v1.1 standard for educational data exchange
    - Discovery API for district and school information
    - Academy API for professional development
    - Webhook system for real-time event notifications

    All APIs are designed to match Schoolday's actual behavior for seamless local development and testing.
  version: 1.0.0
  contact:
    name: Teacher Evaluation Platform Team
    email: dev@teacher-eval.local
  license:
    name: MIT License
servers:
  - url: http://localhost:3006/api
    description: Local development server
  - url: https://faux-schoolday.teacher-eval.local/api
    description: Containerized environment
security:
  - BearerAuth: []
paths:
...
  /discovery/districts:
    get:
      tags: [Discovery]
      summary: List districts
      description: Get list of school districts with filtering and pagination
      parameters:
        - name: state
          in: query
          schema:
            type: string
          example: CA
        - name: search
          in: query
          schema:
            type: string
        - name: include_schools
          in: query
          schema:
            type: boolean
            default: false
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of districts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DistrictListResponse'
  /discovery/districts/{id}:
    get:
      tags: [Discovery]
      summary: Get district details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: District details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DistrictDetailResponse'
        '404':
          description: District not found

  # OneRoster API Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    District:
      type: object
      properties:
        sourcedId:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [district]
        identifier:
          type: string
        status:
          type: string
          enum: [active, tobedeleted]
        dateLastModified:
          type: string
          format: date-time
        geographic_info:
          type: object
          properties:
            state:
              type: string
            region:
              type: string
            county:
              type: string
            timezone:
              type: string
        enrollment_stats:
          type: object
          properties:
            total_students:
              type: integer
            total_teachers:
              type: integer
            total_schools:
              type: integer
    DistrictListResponse:
      type: object
      properties:
        districts:
          type: array
          items:
            $ref: '#/components/schemas/District'
        pagination:
          $ref: '#/components/schemas/Pagination'
    DistrictDetailResponse:
      type: object
      properties:
        district:
          allOf:
            - $ref: '#/components/schemas/District'
            - type: object
              properties:
                schools:
                  type: array
                  items:
                    $ref: '#/components/schemas/School'

    # OneRoster Schemas
...
    School:
      type: object
      properties:
        sourcedId:
          type: string
        name:
          type: string
        schoolType:
          type: string
          enum: [elementary, middle, high, k12]
        address:
          type: string
        phone:
          type: string
        email:
          type: string
    Pagination:
      type: object
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        hasMore:
          type: boolean
...